Подключен и настроен роутинг в приложении
