+++ src/components/About/About.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import CardContent from '@material-ui/core/CardContent';
+import CircularProgress from '@material-ui/core/CircularProgress';
+
+const About = () => (
+	<CardContent>
+		<h1>Обо мне</h1>
+	</CardContent>
+);
+
+export default About;
