+++ src/components/About/About.module.css
@@ -1,0 +1,3 @@
+.content{
+	display: none;
+}
