+++ src/components/App/App.js
@@ -1,92 +1,35 @@
-import React, {useState, useEffect} from 'react';
-import ItemList from '../ItemList/ItemList';
-import InputItem from '../InputItem/InputItem';
-import Footer from '../Footer/Footer';
+import React from 'react';
+import { BrowserRouter as Router, Route, Link } from 'react-router-dom';
+import Card from '@material-ui/core/Card';
+import MenuItem from '@material-ui/core/MenuItem';
+import MenuList from '@material-ui/core/MenuList';
+
+import Todo from '../Todo/Todo';
+import About from '../About/About';
+import Contacts from '../Contacts/Contacts';
+
 import styles from './App.module.css';
 
-const App = () => {
-		const initialState = {
-			items: [
-					{
-						value: 'Написать новое приложение',
-						isDone: true,
-						id:1
-					},
-					{
-						value: 'Написать props-ы',
-						isDone: false,
-						id:2
-					},
-					{
-						value: 'стилизовать',
-						isDone: false,
-						id:3
-					},
-					{
-						value: 'Сделать все дела',
-						isDone: true,
-						id:4
-					}
-				],
-				count: 4
-	};
+const App = () =>
+	(<Router>
+		<div className={styles.wrap}>
 
-	const [items, setItems] = useState (initialState.items);
-	const [count, setCount] = useState (initialState.count);
-	
-	useEffect(() => {console.log('mount');}, []);
-	useEffect(() => {console.log('update');});
-	
-	const onClickDone = id => {
-		const newItemList = items.map(item =>{
-			const newItem = { ...item};
-			if(item.id ===id) {
-				newItem.isDone = !item.isDone;
-			}
-			return newItem
-		});
+			<Card className={styles.sidebar}>
+				<MenuList>
+		          <Link to='/' className={styles.link}><MenuItem>Обо мне</MenuItem></Link>
+		          <Link to='/todo' className={styles.link}><MenuItem>Дела</MenuItem></Link>
+		          <Link to='/contacts' className={styles.link}><MenuItem>Контакты</MenuItem></Link>
+	        </MenuList>
+			</Card>
 
-		setItems(newItemList);
-	};
+			<Card className={styles.content}>
+				<Route path='/' exact component={About} />
+				<Route path='/todo' component={Todo} />
+				<Route path='/contacts' component={Contacts} />
+			</Card>
 
-	const onClickDelete = id => {
-			const newItemList = items.filter(item =>{
-				return item.id !== id; 
-			});
-			setItems(newItemList)
-			setCount(count - 1)
-		};
-
-	const onClickAdd = value => {
-		setItems(
-			[...items,
-				{
-					value: value.toLowerCase(),
-					isDone: false,
-					id: count + 1
-				}]);
-		setCount(count + 1) 
-	}
-
-	const onClickDeleteDone = id => {
-		const newItemList = items.filter(item =>{
-			return item.isDone !== true; 
-		});
-		setItems(newItemList);
-	};
-
-		return (
-			<div className={styles.wrap}>
-				<h1 className={styles.title}>Важные дела:</h1>
-				<InputItem onClickAdd={onClickAdd} />
-				<ItemList 
-					items={items} 
-					onClickDone={onClickDone}  
-					onClickDelete={onClickDelete} />
-					<Footer count={items.length}
-					onClickDeleteDone={onClickDeleteDone}
-				/>
-			</div>);
-}
+		</div>
+	</Router>);
+		
 
 export default App;
