+++ src/components/Contacts/Contacts.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import CardContent from '@material-ui/core/CardContent';
+import CircularProgress from '@material-ui/core/CircularProgress';
+
+const Contacts = () => (
+	<CardContent>
+		<h1>Контакты</h1>
+	</CardContent>
+);
+
+export default Contacts;
